Moving classes from osa_refresh into library cleaner_utils. Updated readme.txt for osa_refresh.py
